pkg/mj: refactor main entrypoint to make it unit-testable
