+++ .gitignore
@@ -1,4 +1,4 @@
-mj
+/mj
 
 # Compiled Object files, Static and Dynamic libs (Shared Objects)
 *.o
