+++ pkg/mj/builder.go
@@ -1,4 +1,4 @@
-package main
+package mj
 
 import (
 	"encoding/json"
