+++ pkg/mj/builder_test.go
@@ -1,4 +1,4 @@
-package main
+package mj
 
 import (
 	"reflect"
