+++ pkg/mj/processor.go
@@ -1,4 +1,4 @@
-package main
+package mj
 
 import (
 	"fmt"
