+++ pkg/mj/version.go
@@ -1,4 +1,4 @@
-package main
+package mj
 
 var BuildCommit = "HEAD"
 var BuildDate = "0000-00-00T00:00:00Z"
